AnomalyDetectionService/requirements.txt
Flask==3.1.0
blinker==1.9.0
h5py==3.12.1
joblib==1.4.2
numpy==2.0.2
pandas==2.2.3
requests==2.32.3
scikit-learn==1.6.0
tensorflow==2.18.0
Werkzeug==3.1.3
keras==3.7.0
sklearn==0.0
kafka-python==2.0.2
